Factor out the libjpeg-decoding parts of test_utils.cc
This is preparation for using jpegli's own jpeg headers for
compiling the jpegli-static library.
