+++ lib/jpegli/common.h
@@ -25,6 +25,8 @@
 #include <jpeglib.h>
 /* clang-format on */
 
+#include "lib/jpegli/types.h"
+
 #if defined(__cplusplus) || defined(c_plusplus)
 extern "C" {
 #endif
@@ -38,27 +40,6 @@ void jpegli_destroy(j_common_ptr cinfo);
 JQUANT_TBL* jpegli_alloc_quant_table(j_common_ptr cinfo);
 
 JHUFF_TBL* jpegli_alloc_huff_table(j_common_ptr cinfo);
-
-//
-// New API structs and functions that are not available in libjpeg
-//
-// NOTE: This part of the API is still experimental and will probably change in
-// the future.
-//
-
-typedef enum {
-  JPEGLI_TYPE_FLOAT = 0,
-  JPEGLI_TYPE_UINT8 = 2,
-  JPEGLI_TYPE_UINT16 = 3,
-} JpegliDataType;
-
-typedef enum {
-  JPEGLI_NATIVE_ENDIAN = 0,
-  JPEGLI_LITTLE_ENDIAN = 1,
-  JPEGLI_BIG_ENDIAN = 2,
-} JpegliEndianness;
-
-int jpegli_bytes_per_sample(JpegliDataType data_type);
 
 #if defined(__cplusplus) || defined(c_plusplus)
 }  // extern "C"
