+++ lib/jpegli/input_suspension_test.cc
@@ -243,7 +243,8 @@ TEST_P(InputSuspensionTestParam, InputOutputLockStepNonBuffered) {
     while (jpegli_read_header(&cinfo, TRUE) == JPEG_SUSPENDED) {
       JXL_CHECK(src.LoadNextChunk());
     }
-    SetDecompressParams(dparams, &cinfo, true);
+    SetDecompressParams(dparams, &cinfo);
+    jpegli_set_output_format(&cinfo, dparams.data_type, dparams.endianness);
     if (config.jparams.add_marker) {
       EXPECT_EQ(num_markers_seen, kMarkerSequenceLen);
       EXPECT_EQ(0, memcmp(markers_seen, kMarkerSequence, num_markers_seen));
@@ -299,7 +300,8 @@ TEST_P(InputSuspensionTestParam, InputOutputLockStepBuffered) {
     while (jpegli_read_header(&cinfo, TRUE) == JPEG_SUSPENDED) {
       JXL_CHECK(src.LoadNextChunk());
     }
-    SetDecompressParams(dparams, &cinfo, true);
+    SetDecompressParams(dparams, &cinfo);
+    jpegli_set_output_format(&cinfo, dparams.data_type, dparams.endianness);
 
     cinfo.buffered_image = TRUE;
     cinfo.raw_data_out = dparams.output_mode == RAW_DATA;
