+++ lib/jpegli/test_utils.cc
@@ -21,6 +21,10 @@
 
 namespace jpegli {
 
+#define JPEG_API_FN(name) jpegli_##name
+#include "lib/jpegli/test_utils-inl.h"
+#undef JPEG_API_FN
+
 #if defined(TEST_DATA_PATH)
 std::string GetTestDataPath(const std::string& filename) {
   return std::string(TEST_DATA_PATH "/") + filename;
@@ -211,7 +215,7 @@ std::ostream& operator<<(std::ostream& os, const TestImage& input) {
   os << input.xsize << "x" << input.ysize;
   os << IOMethodName(input.data_type, input.endianness);
   if (input.color_space != JCS_RGB) {
-    os << "InputColor" << ColorSpaceName(input.color_space);
+    os << "InputColor" << ColorSpaceName((J_COLOR_SPACE)input.color_space);
   }
   if (input.color_space == JCS_UNKNOWN) {
     os << input.components;
@@ -223,7 +227,8 @@ std::ostream& operator<<(std::ostream& os, const CompressParams& jparams) {
   os << "Q" << jparams.quality;
   os << SamplingId(jparams);
   if (jparams.set_jpeg_colorspace) {
-    os << "JpegColor" << ColorSpaceName(jparams.jpeg_color_space);
+    os << "JpegColor"
+       << ColorSpaceName((J_COLOR_SPACE)jparams.jpeg_color_space);
   }
   if (!jparams.comp_ids.empty()) {
     os << "CID";
@@ -406,7 +411,7 @@ void GeneratePixels(TestImage* img) {
   size_t in_stride = xsize * in_bytes_per_pixel;
   size_t x0 = (xsize - img->xsize) / 2;
   size_t y0 = (ysize - img->ysize) / 2;
-  SetNumChannels(img->color_space, &img->components);
+  SetNumChannels((J_COLOR_SPACE)img->color_space, &img->components);
   size_t out_bytes_per_pixel =
       jpegli_bytes_per_sample(img->data_type) * img->components;
   size_t out_stride = img->xsize * out_bytes_per_pixel;
@@ -420,8 +425,9 @@ void GeneratePixels(TestImage* img) {
       size_t x = x0 + ix;
       size_t idx_in = y * in_stride + x * in_bytes_per_pixel;
       size_t idx_out = iy * out_stride + ix * out_bytes_per_pixel;
-      ConvertPixel(&pixels[idx_in], &img->pixels[idx_out], img->color_space,
-                   img->components, img->data_type, swap_endianness);
+      ConvertPixel(&pixels[idx_in], &img->pixels[idx_out],
+                   (J_COLOR_SPACE)img->color_space, img->components,
+                   img->data_type, swap_endianness);
     }
   }
 }
@@ -485,7 +491,7 @@ void EncodeWithJpegli(const TestImage& input, const CompressParams& jparams,
     jpegli_set_progressive_level(cinfo, 0);
   }
   jpegli_set_defaults(cinfo);
-  cinfo->in_color_space = input.color_space;
+  cinfo->in_color_space = (J_COLOR_SPACE)input.color_space;
   jpegli_default_colorspace(cinfo);
   if (jparams.override_JFIF >= 0) {
     cinfo->write_JFIF_header = jparams.override_JFIF;
@@ -494,7 +500,7 @@ void EncodeWithJpegli(const TestImage& input, const CompressParams& jparams,
     cinfo->write_Adobe_marker = jparams.override_Adobe;
   }
   if (jparams.set_jpeg_colorspace) {
-    jpegli_set_colorspace(cinfo, jparams.jpeg_color_space);
+    jpegli_set_colorspace(cinfo, (J_COLOR_SPACE)jparams.jpeg_color_space);
   }
   if (!jparams.comp_ids.empty()) {
     for (int c = 0; c < cinfo->num_components; ++c) {
@@ -672,472 +678,7 @@ bool EncodeWithJpegli(const TestImage& input, const CompressParams& jparams,
   return success;
 }
 
-void SetScanDecompressParams(const DecompressParams& dparams,
-                             j_decompress_ptr cinfo, int scan_number,
-                             bool is_jpegli) {
-  const ScanDecompressParams* sparams = nullptr;
-  for (const auto& sp : dparams.scan_params) {
-    if (scan_number <= sp.max_scan_number) {
-      sparams = &sp;
-      break;
-    }
-  }
-  if (sparams == nullptr) {
-    return;
-  }
-  if (dparams.quantize_colors) {
-    cinfo->dither_mode = sparams->dither_mode;
-    if (sparams->color_quant_mode == CQUANT_1PASS) {
-      cinfo->two_pass_quantize = FALSE;
-      cinfo->colormap = nullptr;
-    } else if (sparams->color_quant_mode == CQUANT_2PASS) {
-      JXL_CHECK(cinfo->out_color_space = JCS_RGB);
-      cinfo->two_pass_quantize = TRUE;
-      cinfo->colormap = nullptr;
-    } else if (sparams->color_quant_mode == CQUANT_EXTERNAL) {
-      JXL_CHECK(cinfo->out_color_space = JCS_RGB);
-      cinfo->two_pass_quantize = FALSE;
-      bool have_colormap = cinfo->colormap != nullptr;
-      cinfo->actual_number_of_colors = kTestColorMapNumColors;
-      cinfo->colormap = (*cinfo->mem->alloc_sarray)(
-          reinterpret_cast<j_common_ptr>(cinfo), JPOOL_IMAGE,
-          cinfo->actual_number_of_colors, 3);
-      jxl::msan::UnpoisonMemory(cinfo->colormap, 3 * sizeof(JSAMPROW));
-      for (int i = 0; i < kTestColorMapNumColors; ++i) {
-        cinfo->colormap[0][i] = (kTestColorMap[i] >> 16) & 0xff;
-        cinfo->colormap[1][i] = (kTestColorMap[i] >> 8) & 0xff;
-        cinfo->colormap[2][i] = (kTestColorMap[i] >> 0) & 0xff;
-      }
-      if (have_colormap) {
-        if (is_jpegli) {
-          jpegli_new_colormap(cinfo);
-        } else {
-          jpeg_new_colormap(cinfo);
-        }
-      }
-    } else if (sparams->color_quant_mode == CQUANT_REUSE) {
-      JXL_CHECK(cinfo->out_color_space = JCS_RGB);
-      JXL_CHECK(cinfo->colormap);
-    }
-  }
-}
-
-void SetDecompressParams(const DecompressParams& dparams,
-                         j_decompress_ptr cinfo, bool is_jpegli) {
-  cinfo->do_block_smoothing = dparams.do_block_smoothing;
-  cinfo->do_fancy_upsampling = dparams.do_fancy_upsampling;
-  if (dparams.output_mode == RAW_DATA) {
-    cinfo->raw_data_out = TRUE;
-  }
-  if (dparams.set_out_color_space) {
-    cinfo->out_color_space = dparams.out_color_space;
-    if (dparams.out_color_space == JCS_UNKNOWN) {
-      cinfo->jpeg_color_space = JCS_UNKNOWN;
-    }
-  }
-  cinfo->scale_num = dparams.scale_num;
-  cinfo->scale_denom = dparams.scale_denom;
-  cinfo->quantize_colors = dparams.quantize_colors;
-  cinfo->desired_number_of_colors = dparams.desired_number_of_colors;
-  if (!dparams.scan_params.empty()) {
-    if (cinfo->buffered_image) {
-      for (const auto& sparams : dparams.scan_params) {
-        if (sparams.color_quant_mode == CQUANT_1PASS) {
-          cinfo->enable_1pass_quant = TRUE;
-        } else if (sparams.color_quant_mode == CQUANT_2PASS) {
-          cinfo->enable_2pass_quant = TRUE;
-        } else if (sparams.color_quant_mode == CQUANT_EXTERNAL) {
-          cinfo->enable_external_quant = TRUE;
-        }
-      }
-      SetScanDecompressParams(dparams, cinfo, 1, is_jpegli);
-    } else {
-      SetScanDecompressParams(dparams, cinfo, kLastScan, is_jpegli);
-    }
-  }
-  if (is_jpegli) {
-    jpegli_set_output_format(cinfo, dparams.data_type, dparams.endianness);
-  }
-}
-
-void CheckMarkerPresent(j_decompress_ptr cinfo, uint8_t marker_type) {
-  bool marker_found = false;
-  for (jpeg_saved_marker_ptr marker = cinfo->marker_list; marker != nullptr;
-       marker = marker->next) {
-    jxl::msan::UnpoisonMemory(marker, sizeof(*marker));
-    jxl::msan::UnpoisonMemory(marker->data, marker->data_length);
-    if (marker->marker == marker_type &&
-        marker->data_length == sizeof(kMarkerData) &&
-        memcmp(marker->data, kMarkerData, sizeof(kMarkerData)) == 0) {
-      marker_found = true;
-    }
-  }
-  JXL_CHECK(marker_found);
-}
-
-void VerifyHeader(const CompressParams& jparams, j_decompress_ptr cinfo) {
-  if (jparams.set_jpeg_colorspace) {
-    JXL_CHECK(cinfo->jpeg_color_space == jparams.jpeg_color_space);
-  }
-  if (jparams.override_JFIF >= 0) {
-    JXL_CHECK(cinfo->saw_JFIF_marker == jparams.override_JFIF);
-  }
-  if (jparams.override_Adobe >= 0) {
-    JXL_CHECK(cinfo->saw_Adobe_marker == jparams.override_Adobe);
-  }
-  if (jparams.add_marker) {
-    CheckMarkerPresent(cinfo, kSpecialMarker0);
-    CheckMarkerPresent(cinfo, kSpecialMarker1);
-  }
-  jxl::msan::UnpoisonMemory(
-      cinfo->comp_info, cinfo->num_components * sizeof(cinfo->comp_info[0]));
-  int max_h_samp_factor = 1;
-  int max_v_samp_factor = 1;
-  for (int i = 0; i < cinfo->num_components; ++i) {
-    jpeg_component_info* comp = &cinfo->comp_info[i];
-    if (!jparams.comp_ids.empty()) {
-      JXL_CHECK(comp->component_id == jparams.comp_ids[i]);
-    }
-    if (!jparams.h_sampling.empty()) {
-      JXL_CHECK(comp->h_samp_factor == jparams.h_sampling[i]);
-    }
-    if (!jparams.v_sampling.empty()) {
-      JXL_CHECK(comp->v_samp_factor == jparams.v_sampling[i]);
-    }
-    if (!jparams.quant_indexes.empty()) {
-      JXL_CHECK(comp->quant_tbl_no == jparams.quant_indexes[i]);
-    }
-    max_h_samp_factor = std::max(max_h_samp_factor, comp->h_samp_factor);
-    max_v_samp_factor = std::max(max_v_samp_factor, comp->v_samp_factor);
-  }
-  JXL_CHECK(max_h_samp_factor == cinfo->max_h_samp_factor);
-  JXL_CHECK(max_v_samp_factor == cinfo->max_v_samp_factor);
-  int referenced_tables[NUM_QUANT_TBLS] = {};
-  for (int i = 0; i < cinfo->num_components; ++i) {
-    jpeg_component_info* comp = &cinfo->comp_info[i];
-    JXL_CHECK(comp->width_in_blocks ==
-              DivCeil(cinfo->image_width * comp->h_samp_factor,
-                      max_h_samp_factor * DCTSIZE));
-    JXL_CHECK(comp->height_in_blocks ==
-              DivCeil(cinfo->image_height * comp->v_samp_factor,
-                      max_v_samp_factor * DCTSIZE));
-    referenced_tables[comp->quant_tbl_no] = 1;
-  }
-  for (const auto& table : jparams.quant_tables) {
-    JQUANT_TBL* quant_table = cinfo->quant_tbl_ptrs[table.slot_idx];
-    if (!referenced_tables[table.slot_idx]) {
-      JXL_CHECK(quant_table == nullptr);
-      continue;
-    }
-    JXL_CHECK(quant_table != nullptr);
-    jxl::msan::UnpoisonMemory(quant_table, sizeof(*quant_table));
-    for (int k = 0; k < DCTSIZE2; ++k) {
-      JXL_CHECK(quant_table->quantval[k] == table.quantval[k]);
-    }
-  }
-}
-
-void VerifyScanHeader(const CompressParams& jparams, j_decompress_ptr cinfo) {
-  JXL_CHECK(cinfo->input_scan_number > 0);
-  if (cinfo->progressive_mode) {
-    JXL_CHECK(cinfo->Ss != 0 || cinfo->Se != 63);
-  } else {
-    JXL_CHECK(cinfo->Ss == 0 && cinfo->Se == 63);
-  }
-  if (jparams.progressive_mode > 2) {
-    JXL_CHECK(jparams.progressive_mode < 3 + kNumTestScripts);
-    const ScanScript& script = kTestScript[jparams.progressive_mode - 3];
-    JXL_CHECK(cinfo->input_scan_number <= script.num_scans);
-    const jpeg_scan_info& scan = script.scans[cinfo->input_scan_number - 1];
-    JXL_CHECK(cinfo->comps_in_scan == scan.comps_in_scan);
-    for (int i = 0; i < cinfo->comps_in_scan; ++i) {
-      JXL_CHECK(cinfo->cur_comp_info[i]->component_index ==
-                scan.component_index[i]);
-    }
-    JXL_CHECK(cinfo->Ss == scan.Ss);
-    JXL_CHECK(cinfo->Se == scan.Se);
-    JXL_CHECK(cinfo->Ah == scan.Ah);
-    JXL_CHECK(cinfo->Al == scan.Al);
-  }
-  if (jparams.restart_interval > 0) {
-    JXL_CHECK(cinfo->restart_interval == jparams.restart_interval);
-  } else if (jparams.restart_in_rows > 0) {
-    JXL_CHECK(cinfo->restart_interval ==
-              jparams.restart_in_rows * cinfo->MCUs_per_row);
-  }
-  if (jparams.progressive_mode == 0 && jparams.optimize_coding == 0) {
-    if (cinfo->jpeg_color_space == JCS_RGB) {
-      JXL_CHECK(cinfo->comp_info[0].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[1].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[2].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[0].ac_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[1].ac_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[2].ac_tbl_no == 0);
-    } else if (cinfo->jpeg_color_space == JCS_YCbCr) {
-      JXL_CHECK(cinfo->comp_info[0].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[1].dc_tbl_no == 1);
-      JXL_CHECK(cinfo->comp_info[2].dc_tbl_no == 1);
-      JXL_CHECK(cinfo->comp_info[0].ac_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[1].ac_tbl_no == 1);
-      JXL_CHECK(cinfo->comp_info[2].ac_tbl_no == 1);
-    } else if (cinfo->jpeg_color_space == JCS_CMYK) {
-      JXL_CHECK(cinfo->comp_info[0].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[1].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[2].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[3].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[0].ac_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[1].ac_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[2].ac_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[3].ac_tbl_no == 0);
-    } else if (cinfo->jpeg_color_space == JCS_YCCK) {
-      JXL_CHECK(cinfo->comp_info[0].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[1].dc_tbl_no == 1);
-      JXL_CHECK(cinfo->comp_info[2].dc_tbl_no == 1);
-      JXL_CHECK(cinfo->comp_info[3].dc_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[0].ac_tbl_no == 0);
-      JXL_CHECK(cinfo->comp_info[1].ac_tbl_no == 1);
-      JXL_CHECK(cinfo->comp_info[2].ac_tbl_no == 1);
-      JXL_CHECK(cinfo->comp_info[3].ac_tbl_no == 0);
-    }
-    if (jparams.use_flat_dc_luma_code) {
-      JHUFF_TBL* tbl = cinfo->dc_huff_tbl_ptrs[0];
-      jxl::msan::UnpoisonMemory(tbl, sizeof(*tbl));
-      for (int i = 0; i < 15; ++i) {
-        JXL_CHECK(tbl->huffval[i] == i);
-      }
-    }
-  }
-}
-
-void UnmapColors(uint8_t* row, size_t xsize, int components,
-                 JSAMPARRAY colormap, size_t num_colors) {
-  JXL_CHECK(colormap != nullptr);
-  std::vector<uint8_t> tmp(xsize * components);
-  for (size_t x = 0; x < xsize; ++x) {
-    JXL_CHECK(row[x] < num_colors);
-    for (int c = 0; c < components; ++c) {
-      tmp[x * components + c] = colormap[c][row[x]];
-    }
-  }
-  memcpy(row, tmp.data(), tmp.size());
-}
-
-void ReadOutputPass(j_decompress_ptr cinfo, const DecompressParams& dparams,
-                    TestImage* output) {
-  JDIMENSION xoffset = 0;
-  JDIMENSION yoffset = 0;
-  JDIMENSION xsize_cropped = cinfo->output_width;
-  JDIMENSION ysize_cropped = cinfo->output_height;
-  if (dparams.crop_output) {
-    xoffset = xsize_cropped = cinfo->output_width / 3;
-    yoffset = ysize_cropped = cinfo->output_height / 3;
-    jpeg_crop_scanline(cinfo, &xoffset, &xsize_cropped);
-    JXL_CHECK(xsize_cropped == cinfo->output_width);
-  }
-  output->xsize = xsize_cropped;
-  output->ysize = ysize_cropped;
-  output->components = cinfo->out_color_components;
-  if (cinfo->quantize_colors) {
-    jxl::msan::UnpoisonMemory(cinfo->colormap, cinfo->out_color_components *
-                                                   sizeof(cinfo->colormap[0]));
-    for (int c = 0; c < cinfo->out_color_components; ++c) {
-      jxl::msan::UnpoisonMemory(
-          cinfo->colormap[c],
-          cinfo->actual_number_of_colors * sizeof(cinfo->colormap[c][0]));
-    }
-  }
-  if (!cinfo->raw_data_out) {
-    size_t stride = output->xsize * output->components;
-    output->pixels.resize(output->ysize * stride);
-    output->color_space = cinfo->out_color_space;
-    if (yoffset > 0) {
-      jpeg_skip_scanlines(cinfo, yoffset);
-    }
-    for (size_t y = 0; y < output->ysize; ++y) {
-      JSAMPROW rows[] = {
-          reinterpret_cast<JSAMPLE*>(&output->pixels[y * stride])};
-      JXL_CHECK(1 == jpeg_read_scanlines(cinfo, rows, 1));
-      jxl::msan::UnpoisonMemory(
-          rows[0], sizeof(JSAMPLE) * cinfo->output_components * output->xsize);
-      if (cinfo->quantize_colors) {
-        UnmapColors(rows[0], cinfo->output_width, cinfo->out_color_components,
-                    cinfo->colormap, cinfo->actual_number_of_colors);
-      }
-    }
-    if (cinfo->output_scanline < cinfo->output_height) {
-      jpeg_skip_scanlines(cinfo, cinfo->output_height - cinfo->output_scanline);
-    }
-  } else {
-    output->color_space = cinfo->jpeg_color_space;
-    for (int c = 0; c < cinfo->num_components; ++c) {
-      size_t xsize = cinfo->comp_info[c].width_in_blocks * DCTSIZE;
-      size_t ysize = cinfo->comp_info[c].height_in_blocks * DCTSIZE;
-      std::vector<uint8_t> plane(ysize * xsize);
-      output->raw_data.emplace_back(std::move(plane));
-    }
-    while (cinfo->output_scanline < cinfo->output_height) {
-      size_t iMCU_height = cinfo->max_v_samp_factor * DCTSIZE;
-      JXL_CHECK(cinfo->output_scanline == cinfo->output_iMCU_row * iMCU_height);
-      std::vector<std::vector<JSAMPROW>> rowdata(cinfo->num_components);
-      std::vector<JSAMPARRAY> data(cinfo->num_components);
-      for (int c = 0; c < cinfo->num_components; ++c) {
-        size_t xsize = cinfo->comp_info[c].width_in_blocks * DCTSIZE;
-        size_t ysize = cinfo->comp_info[c].height_in_blocks * DCTSIZE;
-        size_t num_lines = cinfo->comp_info[c].v_samp_factor * DCTSIZE;
-        rowdata[c].resize(num_lines);
-        size_t y0 = cinfo->output_iMCU_row * num_lines;
-        for (size_t i = 0; i < num_lines; ++i) {
-          rowdata[c][i] =
-              y0 + i < ysize ? &output->raw_data[c][(y0 + i) * xsize] : nullptr;
-        }
-        data[c] = &rowdata[c][0];
-      }
-      JXL_CHECK(iMCU_height ==
-                jpeg_read_raw_data(cinfo, &data[0], iMCU_height));
-    }
-  }
-  JXL_CHECK(cinfo->total_iMCU_rows ==
-            DivCeil(cinfo->image_height, cinfo->max_v_samp_factor * DCTSIZE));
-}
-
-void CopyCoefficients(j_decompress_ptr cinfo, jvirt_barray_ptr* coef_arrays,
-                      TestImage* output) {
-  output->xsize = cinfo->image_width;
-  output->ysize = cinfo->image_height;
-  output->components = cinfo->num_components;
-  output->color_space = cinfo->out_color_space;
-  j_common_ptr comptr = reinterpret_cast<j_common_ptr>(cinfo);
-  for (int c = 0; c < cinfo->num_components; ++c) {
-    jpeg_component_info* comp = &cinfo->comp_info[c];
-    std::vector<JCOEF> coeffs(comp->width_in_blocks * comp->height_in_blocks *
-                              DCTSIZE2);
-    for (size_t by = 0; by < comp->height_in_blocks; ++by) {
-      JBLOCKARRAY ba = (*cinfo->mem->access_virt_barray)(comptr, coef_arrays[c],
-                                                         by, 1, true);
-      size_t stride = comp->width_in_blocks * sizeof(JBLOCK);
-      size_t offset = by * comp->width_in_blocks * DCTSIZE2;
-      memcpy(&coeffs[offset], ba[0], stride);
-    }
-    output->coeffs.emplace_back(std::move(coeffs));
-  }
-}
-
-// Verifies that an image encoded with libjpegli can be decoded with libjpeg,
-// and checks that the jpeg coding metadata matches jparams.
-void DecodeAllScansWithLibjpeg(const CompressParams& jparams,
-                               const DecompressParams& dparams,
-                               const std::vector<uint8_t>& compressed,
-                               std::vector<TestImage>* output_progression) {
-  jpeg_decompress_struct cinfo = {};
-  const auto try_catch_block = [&]() {
-    ERROR_HANDLER_SETUP(jpeg);
-    jpeg_create_decompress(&cinfo);
-    jpeg_mem_src(&cinfo, compressed.data(), compressed.size());
-    if (jparams.add_marker) {
-      jpeg_save_markers(&cinfo, kSpecialMarker0, 0xffff);
-      jpeg_save_markers(&cinfo, kSpecialMarker1, 0xffff);
-    }
-    JXL_CHECK(JPEG_REACHED_SOS ==
-              jpeg_read_header(&cinfo, /*require_image=*/TRUE));
-    cinfo.buffered_image = TRUE;
-    SetDecompressParams(dparams, &cinfo, /*is_jpegli=*/false);
-    VerifyHeader(jparams, &cinfo);
-    JXL_CHECK(jpeg_start_decompress(&cinfo));
-    // start decompress should not read the whole input in buffered image mode
-    JXL_CHECK(!jpeg_input_complete(&cinfo));
-    JXL_CHECK(cinfo.output_scan_number == 0);
-    int sos_marker_cnt = 1;  // read header reads the first SOS marker
-    while (!jpeg_input_complete(&cinfo)) {
-      JXL_CHECK(cinfo.input_scan_number == sos_marker_cnt);
-      if (dparams.skip_scans && (cinfo.input_scan_number % 2) != 1) {
-        int result = JPEG_SUSPENDED;
-        while (result != JPEG_REACHED_SOS && result != JPEG_REACHED_EOI) {
-          result = jpeg_consume_input(&cinfo);
-        }
-        if (result == JPEG_REACHED_SOS) ++sos_marker_cnt;
-        continue;
-      }
-      SetScanDecompressParams(dparams, &cinfo, cinfo.input_scan_number,
-                              /*is_jpegli=*/false);
-      JXL_CHECK(jpeg_start_output(&cinfo, cinfo.input_scan_number));
-      // start output sets output_scan_number, but does not change
-      // input_scan_number
-      JXL_CHECK(cinfo.output_scan_number == cinfo.input_scan_number);
-      JXL_CHECK(cinfo.input_scan_number == sos_marker_cnt);
-      VerifyScanHeader(jparams, &cinfo);
-      TestImage output;
-      ReadOutputPass(&cinfo, dparams, &output);
-      output_progression->emplace_back(std::move(output));
-      // read scanlines/read raw data does not change input/output scan number
-      if (!cinfo.progressive_mode) {
-        JXL_CHECK(cinfo.input_scan_number == sos_marker_cnt);
-        JXL_CHECK(cinfo.output_scan_number == cinfo.input_scan_number);
-      }
-      JXL_CHECK(jpeg_finish_output(&cinfo));
-      ++sos_marker_cnt;  // finish output reads the next SOS marker or EOI
-      if (dparams.output_mode == COEFFICIENTS) {
-        jvirt_barray_ptr* coef_arrays = jpeg_read_coefficients(&cinfo);
-        JXL_CHECK(coef_arrays != nullptr);
-        CopyCoefficients(&cinfo, coef_arrays, &output_progression->back());
-      }
-    }
-    JXL_CHECK(jpeg_finish_decompress(&cinfo));
-    return true;
-  };
-  JXL_CHECK(try_catch_block());
-  jpeg_destroy_decompress(&cinfo);
-}
-
-void DecodeWithLibjpeg(const CompressParams& jparams,
-                       const DecompressParams& dparams, j_decompress_ptr cinfo,
-                       TestImage* output) {
-  if (jparams.add_marker) {
-    jpeg_save_markers(cinfo, kSpecialMarker0, 0xffff);
-    jpeg_save_markers(cinfo, kSpecialMarker1, 0xffff);
-  }
-  if (!jparams.icc.empty()) {
-    jpeg_save_markers(cinfo, JPEG_APP0 + 2, 0xffff);
-  }
-  JXL_CHECK(JPEG_REACHED_SOS ==
-            jpeg_read_header(cinfo, /*require_image=*/TRUE));
-  if (!jparams.icc.empty()) {
-    uint8_t* icc_data = nullptr;
-    unsigned int icc_len;
-    JXL_CHECK(jpeg_read_icc_profile(cinfo, &icc_data, &icc_len));
-    JXL_CHECK(icc_data);
-    jxl::msan::UnpoisonMemory(icc_data, icc_len);
-    JXL_CHECK(0 == memcmp(jparams.icc.data(), icc_data, icc_len));
-    free(icc_data);
-  }
-  SetDecompressParams(dparams, cinfo, /*is_jpegli=*/false);
-  VerifyHeader(jparams, cinfo);
-  if (dparams.output_mode == COEFFICIENTS) {
-    jvirt_barray_ptr* coef_arrays = jpeg_read_coefficients(cinfo);
-    JXL_CHECK(coef_arrays != nullptr);
-    CopyCoefficients(cinfo, coef_arrays, output);
-  } else {
-    JXL_CHECK(jpeg_start_decompress(cinfo));
-    VerifyScanHeader(jparams, cinfo);
-    ReadOutputPass(cinfo, dparams, output);
-  }
-  JXL_CHECK(jpeg_finish_decompress(cinfo));
-}
-
-void DecodeWithLibjpeg(const CompressParams& jparams,
-                       const DecompressParams& dparams,
-                       const std::vector<uint8_t>& compressed,
-                       TestImage* output) {
-  jpeg_decompress_struct cinfo = {};
-  const auto try_catch_block = [&]() {
-    ERROR_HANDLER_SETUP(jpeg);
-    jpeg_create_decompress(&cinfo);
-    jpeg_mem_src(&cinfo, compressed.data(), compressed.size());
-    DecodeWithLibjpeg(jparams, dparams, &cinfo, output);
-    return true;
-  };
-  JXL_CHECK(try_catch_block());
-  jpeg_destroy_decompress(&cinfo);
-}
+int NumTestScanScripts() { return kNumTestScripts; }
 
 void DumpImage(const TestImage& image, const std::string fn) {
   JXL_CHECK(image.components == 1 || image.components == 3);
